Calculator Features, Fixes, Accessibility
Added new features to the calculator page.

The user can now mark a row for attention - useful as a reminder for corrections, etc.
The attention toggle is transferrable when copying the calculators.

The row and calculator reset show the soon-to-be-removed elements more explicitly.
The box shadow of the elements changes its color when hovering over the reset buttons.

Fixed an issue with the multiplier when copying the calculator ("stop" functionality).
Previously, when copying the calculator with a multiplier set to 0 or 10 (min/max values), the "-" and "+" buttons would not be disabled on hover (the pointer cursor would still be visible), which is now fixed.

Improved accessibility by adding aria-labels to the input fields to enhance the UX (calculator page & navigation).