Calculator Combination, Design Changes, Bug Fixes
Implemented a feature for calculating the total amount of nutritional properties from the chosen calculators.
Everything is placed in the sticky container on the side of the page - below the calculator links.
The user can choose which calculators should be taken into account.
The calculated data consists of weight ratio and calorie percentage. Both are displayed as text and on a graph.

Fixed a bug related to the "current calculator highlight" on the calculator page.
Previously, when the user clicked the "scroll-to-top" button, it was not properly registered, and the highlighted section would not change to the top one.

Changed the multiplier box shadow to the "attention color" when the row is marked for attention (the color of the text still corresponds to the value).

Changed the Dark Mode implementation to disable all transitions for a brief moment when changing the theme.
This leads to a significantly improved, smoother theme change.
Previously, a theme switch would cause weird visual glitches and a choppy animation.

Slightly modified the on-load background animation for a smoother appearance.

Implemented other minor design changes.

Modified the readme file.