v1.2.1 - Nav & Footer Changes
Modified the footer HTML code to be inserted via a JS script.
This will result in easier future changes, if necessary.
Changed the year in the footer credits to be set manually and added the cookbook version info.

Modified the navigation JS insertion code to use "appendChild" instead of "innerHTML" to avoid redrawing the entire page.
This change should enhance the user experience and adhere to better coding practices.

Changed the weight value in the total calculator nutritional values to be rounded up to one decimal place.