v1.3.0 - Calculator Item Search
Added the food item search (powered by the Nutritionix API).
The user can choose which macronutrients should be inserted into the calculator inputs (in the calculator settings):
- calories
- fat
- carbs
- dietary fiber
- protein

A list of results will be displayed after the user:
- enabled "auto-search" option in the settings and types something in the name input
- clicks the "search button"

The typing speed is throttled to prevent excessive fetch requests when typing quickly or holding a key.

There is an icon to indicate if the input value is validated.
It appears dimmed if not validated and flashes red when it's clicked & not validated.
When the fetch request is pending, the icon changes to an animated loading symbol.

If there are any errors, they are displayed instead of the item list.

The user can click on an item from the list and the macronutrient values will be inserted into the calculator inputs.
The results, values and localStorage are handled by dispatching a custom event.

Now, the calculator name is removed when the calculator is reset.

The last modification date of calculator values is saved to the localStorage.

Changed the calculator focus navigation to accommodate for the new "name" input structure.

Implemented minor design changes.

Updated the readme file.