v1.3.3 - Calculator Merging Fix
Fixed a bug where if the user entered a value in the calculator input, the updated calculated result would not display correctly in the Total Results section.
This issue occurred when loading the browser for the first time.

Now, after the value is calculated the script sends a custom event to the Total Results code to trigger calculation instead of relying on the input event listener.